requirements.txt
flask
torch
transformers
pillow
ultralytics